requirements.txt
anyjson==0.3.3
billiard==3.3.0.21
celery==3.1.19
Django==1.8.3
django-celery==3.1.17
django-model-utils==2.3.1
djangorestframework==3.2.3
mysqlclient==1.3.6
Pillow==2.9.0
psycopg2==2.6.1
pypng==0.0.18
PyQRCode==1.1
pytz==2015.7
six==1.9.0
wheel==0.24.0
django-cors-headers==1.1.0
djangorestframework-jwt==1.8.0
django-filter==0.13.0
mkdocs==0.15.3